Splitted manipulate knowledge component
